devtools/client/webide/test/test_device_preferences.html

<!DOCTYPE html>

<html>

  <head>
    <meta charset="utf8">
    <title></title>
    <script src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>
    <script src="chrome://mochikit/content/chrome-harness.js"></script>
    <script type="application/javascript" src="head.js"></script>
    <script type="application/javascript" src="device_front_shared.js"></script>
    <link rel="stylesheet" type="text/css" href="chrome://mochikit/content/tests/SimpleTest/test.css">
  </head>

  <body>

    <script type="application/javascript">
      window.onload = function() {
        SimpleTest.waitForExplicitFinish();

        (async function() {
          DebuggerServer.init();
          DebuggerServer.registerAllActors();

          const win = await openWebIDE();

          const prefIframe = win.document.querySelector("#deck-panel-devicepreferences");
          const docRuntime = getRuntimeDocument(win);

          win.AppManager.update("runtime-list");

          await connectToLocalRuntime(win);

          const prefs = docRuntime.querySelector("#runtime-preferences");

          ok(!prefs.hasAttribute("disabled"), "device prefs cmd enabled");

          const deck = win.document.querySelector("#deck");

          win.Cmds.showDevicePrefs();
          is(deck.selectedPanel, prefIframe, "device preferences iframe selected");

          await nextTick();

          await lazyIframeIsLoaded(prefIframe);

          await prefIframe.contentWindow.getAllPrefs;

          setDocument(prefIframe);

          const fields = doc.querySelectorAll(".editable");

          addNewField();

          const preference = "accessibility.accesskeycausesactivation";

          fieldChange(fields, preference);

          addNewFieldWithEnter();

          editExistingField();

          addNewFieldInteger();

          await editFieldInteger();

          await resetExistingField("accessibility.accesskeycausesactivation");

          addNewFieldBoolean();

          searchFields(deck, "debugger");

          DebuggerServer.destroy();

          await closeWebIDE(win);

          SimpleTest.finish();
        })().catch(e => {
          ok(false, "Exception: " + e);
          SimpleTest.finish();
        });
      };
    </script>
  </body>
</html>
